Remove binary files as well (was skipping them previously). Inspired by lbt

+++ rename-to-my-project.sh
@@ -20,17 +20,28 @@ if [[ $newname != harbour-* ]]; then
 fi
 
 echo Replacing "harbour-helloworld-pro-sailfish" with "$newname"
-# iterating over all files except for the binary ones
-for fl in `find . -name .git -prune -o -type f -print | xargs file | grep ASCII | cut -d: -f1`
+
+# First rename files
+for fl in $(find .  -name .git -prune -o -type f -name 'harbour-helloworld-pro-sailfish*' -print); do
+    mv $fl ${fl/harbour-helloworld-pro-sailfish/$newname}
+    echo Renamed $fl to ${fl/harbour-helloworld-pro-sailfish/$newname}
+done
+
+# Then do substitutions in non-binary files only
+# This magic find line is iterating over non-binary files
+for fl in `find . -type f -print | xargs file | grep ASCII | cut -d: -f1`
 do
     # Ignore this particular file and everything inside .git dir
-    if [[ "$fl" =~ "rename-to-my-project.sh" ]]; then
-        continue
+    if [[ "$fl" =~ "rename-to-my-project.sh" || "$fl" =~ ".git/" ]]
+    then
+         continue
     fi
-    echo Checking $fl
-    futurefl=${fl//harbour-helloworld-pro-sailfish/$newname}
+
     mv $fl $fl.old
-    sed "s/harbour-helloworld-pro-sailfish/$newname/g" $fl.old > $futurefl
+
+    sed "s/harbour-helloworld-pro-sailfish/$newname/g" $fl.old > $fl
     rm -f $fl.old
+    echo Updated $fl
 done
+
 echo Done. Enjoy your $newname app!
